• 1.39j, 14 November 2007       □ Image>Duplicate preserves non-rectangular area selections.       □ Thanks to a patch by Johannes Schindelin, the macro language now         accepts numbers in scientific notation.       □ Plugins>Compile and Run uses the -Xlint:unchecked option on Java 1.5         and later.       □ Fixed a bug that prevented Image>Adjust>Brightness/Contrast, Window/         Level and Color Balance from opening if any of these tools was already         open.       □ Fixed a bug that caused the Roi Manager to sometimes not work as         expected when more than 9999 ROIs were added to it by the particle         analyzer.       □ Fixed several ImageListener interface problems.       □ Fixed a bug that caused File>New>Image to display an invalid message         in the status bar when opening stacks larger than 2GB.       □ Fixed a bug that caused Plugins>Utilities>Memory Monitor to not work         correctly when using more than 2GB.

+++ ij/gui/NewImage.java
@@ -43,7 +43,7 @@ public class NewImage {
 		long bytesPerPixel = 1;
 		if (type==GRAY16) bytesPerPixel = 2;
 		else if (type==GRAY32||type==RGB) bytesPerPixel = 4;
-		long size = (width*height*nSlices*bytesPerPixel);
+		long size = (long)width*height*nSlices*bytesPerPixel;
 		String size2 = size/(1024*1024)+"MB ("+width+"x"+height+"x"+nSlices+")";
 		if ((options&CHECK_AVAILABLE_MEMORY)!=0) {
 			long max = IJ.maxMemory(); // - 100*1024*1024;
